refactor: Restructure Graph with internal stores and proper event namespacing
- Split Graph into public API + internal stores:
  - InternalProcessorLinkGraph: petgraph topology (nodes/edges)
  - InternalProcessorLinkGraphEcsExtension: hecs ECS components
  - Graph: unified public API delegating to both stores

- Rename events to proper namespaces by origin:
  - Runtime*: emitted by StreamRuntime (add/remove/connect/disconnect)
  - Compiler*: emitted by Compiler (create/destroy/wire/unwire)
  - Graph*: reserved for topology changes

- Add Will/Did pattern for all Runtime events
- Add PendingDeletion component for soft-delete mechanism
- Remove dead code (internal_graph accessor, commit_mode methods)
- Update tests and examples to use new APIs
