+++ examples/graph-json-demo/src/main.rs
@@ -41,8 +41,8 @@ fn main() -> Result<()> {
     )?;
 
     // Print the graph as JSON
-    let property_graph = runtime.graph().read();
-    let json = property_graph.graph().read().to_json();
+    let graph = runtime.graph().read();
+    let json = graph.to_json();
     let json_str = serde_json::to_string_pretty(&json).expect("Failed to serialize graph");
 
     println!("{}", json_str);
